[ENH] Allow limits based on stream epoch durations
Allow configurable limits based on stream epoch durations. Limits are
applied to routed stream epochs and thus configurable for all
federated resources implemented. Durations can be either limited based
on the duration of a single stream epoch or based on the total duration
of all routed stream epochs.

Corresponding configuration parameters are:

  - max_stream_epoch_duration
  - max_total_stream_epoch_duration

In case limits are exeeded, the service will respond with HTTP status
code 413 (i.e. in accordance with the FDSN specifications).