[ENH] Implement channel bundeling
Implement channel bundeling for fdsnws-station-xml. Stream epochs are
bundeled such that multiple channels of a unique net.sta.loc.start.end
combination are requested from an endpoint within a single request.

Depending on the implementation and in particular the deployment of the
endpoint webservices of a EIDA DC the algorithm has a performance gain
of roughly up to a factor 10. Note that this number is valid only if no
backend caching is involved (w.r.t. eida-federator).

Note also that the algorithm takes into account that most seismometers
consist of 3 channel components. Thus, if the seismometer is deployed
the all 3 channels belong to a single net.sta.loc.start.end combination.