Start implementing fdsnws-dataselect resource
This is a WIP commit. Still the processor implementation is missing.  At
the time being I'm not sure if it is required to make eidaws-federator
buffering responses in temporary files in order to avoid the roundtrip
to the filesystem.

Though, for sure if buffering on the filesystem is going to be
implemented, I need to find a solution handling (blocking) file I/O.
aiofiles uses a ThreadPoolExecutor for file I/O. A single executor could
be used during the lifetime of a _DataselectAsyncWorker.