[ENH] Dynamically determine the record length
Determine the MiniSEED record length dynamically. When splitting and
aligning (which includes merging potentially occurring overlaps), a
worker assumes the MiniSEED data to be shipped with a uniform record
lenth (with respect to an initially requested stream epoch). I.e. the
record length is determined for the first chunk successfully downloaded
while the subsequent chunks are assumend to be shipped with the same
record length.

The code for determining the record length is strongly based on the
SeisComP3's fdsnws_fetch utility.