[MAINT] Modify method signature and catch OSError
Do not pass the executor to _write* methods. Instead, pass an append
flag when writing the buffer to the response.

When seeking a file fd.seek(arg) raises an OSError in case arg is
smaller than zero. Note, that io.BytesIO does not raise an exception
such that the file descriptor is set to the beginning of the file.