[BUG] Use devel mode to install namespace packages
Since the installation of multiple PEP420 namespace packages apparently
is not possible within a single pip run, install the packages in
development mode (i.e. using the -e flag).

Apparently, in the past there was an open issue w.r.t. the installation
of namespace packages by means of pip:

  - https://github.com/pypa/pip/issues/3