[MAINT] Handle RuntimeError
With the current approach it seems quite tricky achieving both worker
tasks to be cancelled automatically in case the connection is reset by a
peer and to share a web.StreamResponse object between multiple worker
tasks.

The aiohttp docs describe in detail how to prevent tasks from being
cancelled. However they do not describe a scenario as required for a
federating application. See also:

  https://docs.aiohttp.org/en/stable/web_advanced.html

The implementation of something similiar to an asyncio.Pool could solve
the issue. I.e. instead of writing the response straight to the
web.StreamResponse object worker tasks would write endpoint responses to
a queue. These results would be consumend by a coro which is also owner
of the web.StreamResponse.